backend/pptx-generator/requirements.txt
fastapi
uvicorn
httpx
python-pptx
python-multipart